requirements.txt
google-genai=1.16.1
google-generativeai=0.8.5
python-dotenv=1.1.0
streamlit=1.45.1